build: add DSM version as the leading digits.
As of May 2023, Synology requires that the DSM version
be the first two digits of the build number. This is
required to be "60" for DSM6 and "70" for DSM7.

We now  generate package names of the form:
- tailscale-x86_64-1.41.17-600041017-dsm6.spk
- tailscale-x86_64-1.41.17-700041017-dsm7.spk

where:
- "60" or "70" are the DSM version.
- "0" is the major release - 1. Our major release is currently 1.
- "041" is the minor release number, Tailscale 1.41 in this example.
- "017" is the patch number, 17 commits past 1.41.0.

We're keeping the -dsm# at the end of the filename, because
the leading "60" or "70" is subtle and not clear to people who
only occasionally interact with Synology software package formats.

Fixes https://github.com/tailscale/corp/issues/10405

Signed-off-by: Denton Gentry <dgentry@tailscale.com>