start-stop-status: find the correct PID for tailscaled.
Because we pipe tailscaled's output to sed:
`tailscaled | sed -e "..." >> logfile`
Using "$!" retrieves the PID of sed, not the PID of tailscaled.
This breaks the stop action of start-stop-status, it kills the
wrong process.

Use `jobs -p` instead to retrieve the PID of the most recent
process group leader.

Fixes https://github.com/tailscale/corp/issues/10404

Signed-off-by: Denton Gentry <dgentry@tailscale.com>